requirements.txt
requests>=2.28.0
boto3>=1.26.0
pandas>=1.5.0
pyarrow>=12.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
snowflake-connector-python>=3.0.0
